requirements.txt
Flask==3.1.0
paramiko==3.5.1
pywebview==5.4
PyQt6==6.9.0
PyQt6-Qt6==6.9.0
PyQt6-WebEngine==6.9.0
PyQt6-WebEngine-Qt6==6.9.0
PyQt6-sip==13.10.0
QtPy==2.4.3
